refactor organize lib
