+++ frontend/src/lib/utils/parseGameCode.ts
